requirements.txt
Django
gunicorn
django-heroku
social-auth-app-django
django-bootstrap4
Pillow
django-crispy-forms
psycopg2
dj_database_url
python-dotenv
django-bootstrap-modal-forms
django-jquery
django-tables2
django-paypal
django_google_maps
